+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [new version][unreleased]
 
+## Fixed
+
+* Cursor fixes for `Konva.Transformer`
+* Fixed lineHeight behavior for `Konva.Text`
+* Some performance optimizations for `Konva.Text`
+
 ## [2.0.3][2018-04-21]
 
 ### Added